refactor: RowTransformMut (&mut self), move CMS plugin to RowConverter
Adds RowTransformMut trait with &mut self, eliminating the Mutex that
LiteTransform needed to bridge RowTransform's &self with RowConverter's
&mut self. The deprecated RowTransform stays for backward compat with
the old ColorManagement API; LiteTransform keeps its Mutex only for
that path.

PluggableCms::build_source_transform now returns Box<dyn RowTransformMut>
instead of Arc<dyn RowTransform>. The external transform is stored on
RowConverter (not inside ConvertStep), so ConvertStep no longer needs
the ExternalTransform variant or the Arc. This also means ConvertStep
could regain Copy in the future if needed.

ConvertPlan::new_explicit_with_cms is removed — the CMS try-plugin
logic lives on RowConverter::new_explicit_with_cms, which stores the
plugin transform alongside the plan. convert_row checks the external
transform first; if absent, falls through to the plan steps.
